refactor(ask): extract IPC relay and tool registration from index.ts
